멤버와 좋아요 아이템 lazyinitializationerrror 해결

+++ src/main/java/ShopProject/myShopProject/web/Controller/MemberController.java
@@ -36,8 +36,15 @@ public class MemberController {
 
     //회원 탈퇴
     @PostMapping(value = "/member/withdraw")
-    public String removeMember(@RequestParam("memberId") Long memberId ,HttpServletRequest request) {
-        Member loginMember = memberService.findOne(memberId);
+    public String removeMember(@SessionAttribute(name = "loginMember") Member loginMember
+            ,HttpServletRequest request) {
+
+        //카카오 로그인인 경우 카카오 전용 탈퇴
+        if (loginMember.getPassword().equals("K")) {
+            log.info("카카오 로그인, 카카오 전용 탈퇴 컨트롤러 호출");
+            return "redirect:/login/withdrawByKakao";
+        }
+
         memberService.removeMember(loginMember);
         HttpSession session = request.getSession();
         session.invalidate();
